added useTwitterLocationInfo hook
