+++ src/application/agora/app.tsx
@@ -1,24 +1,14 @@
-import { useLocation } from 'react-use';
-
 import { useExtensionSettings } from 'shared/extension';
 import { ErrorBoundary } from 'shared/monitoring';
-import {
-  extractTwitterHandleFromPathname,
-  isTwitterHandlePathname,
-  isTwitterHomePathname,
-  isTwitterHostname,
-} from 'host/twitter';
+import { extractTwitterHandleFromPathname } from 'host/twitter';
+import { useTwitterLocationInfo } from 'host/twitter/hooks';
 
 import { ProposalHandleContainer, ProposalMainContainer } from './widgets';
 
 export const App = () => {
-  const location = useLocation();
   const { experimentalFeatures } = useExtensionSettings();
-
-  // TODO: hide these behind facade
-  const isTwitter = isTwitterHostname(location.hostname ?? '');
-  const isTwitterHandlePage = isTwitterHandlePathname(location.pathname ?? '');
-  const isTwitterHomePage = isTwitterHomePathname(location.pathname ?? '');
+  const { isTwitter, isTwitterHandlePage, isTwitterHomePage } =
+    useTwitterLocationInfo();
 
   if (!experimentalFeatures || !isTwitter) {
     return null;
